Fix scaling bug
The previous logic was pretty broken, resulting in wrong dimensions
in the common case in which height was overly large even after scaling width.

Switch to correct logic where we find the minimum scale factor alpha for which
alpha * width < max_width and alpha * height < max_height. We can find this
just by taking the min of max_width / width and max_height / height in the case
that at least one of the dimensions is overly large. We know that either of these
would yield the appropriate max within that dimension; by taking the min scaling
factor, we know that both will then be in bounds.

The logical flaw should've been obvious, but the lack of unit tests helped it
slip through. TODO: add unit tests.